pyproject.toml
[project]
name = "ansible-mercator"
version = "0.1.0"
description = "A CLI tool that ensures your Ansible variables are defined where they should be."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ansible>=13.1.0",
    "click>=8.3.1",
    "pathlib>=1.0.1",
    "pyyaml>=6.0.3",
]

[project.scripts]
ansible-mercator = "ansible_mercator.__main__:main"

[tool.uv]
package = true
